+++ packages/shared/src/atoms/index.ts
@@ -1,4 +1,9 @@
-export { userAtom, isAuthenticatedAtom, isAuthLoadingAtom } from './auth.atom'
+export {
+  userAtom,
+  isAuthenticatedAtom,
+  isAuthLoadingAtom,
+  isHttpClientReadyAtom,
+} from './auth.atom'
 export { isLoadingAtom } from './ui.atom'
 export { modalConfigAtom } from './modal.atom'
 export type {