+++ packages/shared/src/index.native.ts
@@ -28,6 +28,7 @@ export {
   userAtom,
   isAuthenticatedAtom,
   isAuthLoadingAtom,
+  isHttpClientReadyAtom,
   isLoadingAtom,
   modalConfigAtom,
 } from './atoms'