+++ packages/shared/src/index.ts
@@ -31,6 +31,7 @@ export {
   userAtom,
   isAuthenticatedAtom,
   isAuthLoadingAtom,
+  isHttpClientReadyAtom,
   isLoadingAtom,
   modalConfigAtom,
 } from './atoms'